+++ app/src/main/java/com/sanshy/buysellinventory/UndoSideExp.java
@@ -92,6 +92,7 @@ public class UndoSideExp extends AppCompatActivity {
                                 @Override
                                 public void onClick(DialogInterface dialogInterface, int i) {
                                     MyProgressBar.ShowProgress(UndoSideExp.this);
+tempo.clear();
                                     DatabaseReference mExp = mRootRef.child(userIdMainStatic+"/SideBusiness/Expenditure/"+EidS);
                                     mExp.removeValue();
                                     final DatabaseReference mSideStatementRef = mRootRef.child(userIdMainStatic+"/Statement/SideBStatement/"+DateS);
@@ -118,11 +119,13 @@ public class UndoSideExp extends AppCompatActivity {
                                                 mSideStatementRef.child(PROFIT).setValue(SaveProfit);
                                             }
                                             MyProgressBar.HideProgress();
+tempo.add(1);
                                         }
 
                                         @Override
                                         public void onCancelled(@NonNull DatabaseError databaseError) {
                                             MyProgressBar.HideProgress();
+tempo.add(1);
                                         }
                                     });
                                 }
@@ -143,6 +146,7 @@ public class UndoSideExp extends AppCompatActivity {
 
         Query query = mExp.limitToLast(50);
         MyProgressBar.ShowProgress(this);
+tempo.clear();
         query.addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
@@ -187,6 +191,7 @@ public class UndoSideExp extends AppCompatActivity {
                 remainAmount.setText("Total Money : "+total);
                 
                 MyProgressBar.HideProgress();
+tempo.add(1);
             }
 
             @Override
@@ -223,6 +228,7 @@ public class UndoSideExp extends AppCompatActivity {
 
                         Query query = mSearchRef.orderByChild("remark").equalTo(suggestion_box4.getText().toString());
                         MyProgressBar.ShowProgress(UndoSideExp.this);
+tempo.clear();
                         query.addValueEventListener(new ValueEventListener() {
                             @Override
                             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
@@ -279,11 +285,13 @@ public class UndoSideExp extends AppCompatActivity {
                                 remainAmount.setText("Total Money : "+total);
                                 progressBar.setVisibility(View.INVISIBLE);
                                 MyProgressBar.HideProgress();
+tempo.add(1);
                             }
 
                             @Override
                             public void onCancelled(@NonNull DatabaseError databaseError) {
                                 MyProgressBar.HideProgress();
+tempo.add(1);
                             }
                         });
 
@@ -407,6 +415,7 @@ public class UndoSideExp extends AppCompatActivity {
                 money.clear();
                 Eid.clear();
                 MyProgressBar.ShowProgress(this);
+tempo.clear();
                 for (int i = 0; i < Dates.size(); i++)
                 {
                     final int iFinal = i;
@@ -452,12 +461,14 @@ public class UndoSideExp extends AppCompatActivity {
 
                             if (iFinal==(Dates.size()-1)){
                                 MyProgressBar.HideProgress();
+tempo.add(1);
                             }
                         }
 
                         @Override
                         public void onCancelled(@NonNull DatabaseError databaseError) {
                             MyProgressBar.HideProgress();
+tempo.add(1);
                         }
                     });
 
@@ -508,6 +519,7 @@ public class UndoSideExp extends AppCompatActivity {
                 money.clear();
                 Eid.clear();
                 MyProgressBar.ShowProgress(this);
+tempo.clear();
                 for (int i = 0; i < Dates.size(); i++)
                 {
                     final int iFinal = i;
@@ -553,6 +565,7 @@ public class UndoSideExp extends AppCompatActivity {
                             
                             if (iFinal==(Dates.size()-1)){
                                 MyProgressBar.HideProgress();
+tempo.add(1);
                             }
 
                         }
@@ -560,6 +573,7 @@ public class UndoSideExp extends AppCompatActivity {
                         @Override
                         public void onCancelled(@NonNull DatabaseError databaseError) {
                             MyProgressBar.HideProgress();
+tempo.add(1);
                         }
                     });
 
@@ -574,11 +588,14 @@ public class UndoSideExp extends AppCompatActivity {
 
         progressBar.setVisibility(View.INVISIBLE);
     }
+    ArrayList<Integer> tempo = new ArrayList<>();
     @Override
-    public void onPause() {
+    protected void onPause() {
         super.onPause();
 
-        android.os.Process.killProcess(android.os.Process.myPid());
+        if (tempo.size()==0){
+            android.os.Process.killProcess(android.os.Process.myPid());
+        }
     }
